handle webgl context errors in Editor

+++ editor/public/js/editor/gizmo/gizmo-light.js
@@ -32,6 +32,8 @@ editor.once('load', function () {
     }
     // update lines
     Gizmo.prototype.update = function() {
+        if (! app) return; // webgl not available
+
         if (! this._link || ! this._link.entity)
             return;
 
@@ -118,6 +120,8 @@ editor.once('load', function () {
     };
     // link to entity
     Gizmo.prototype.link = function(obj) {
+        if (! app) return; // webgl not available
+
         this.unlink();
         this._link = obj;
 
@@ -138,6 +142,8 @@ editor.once('load', function () {
     };
     // unlink
     Gizmo.prototype.unlink = function() {
+        if (! app) return; // webgl not available
+
         if (! this._link)
             return;
 
@@ -209,6 +215,7 @@ editor.once('load', function () {
 
     editor.once('viewport:load', function() {
         app = editor.call('viewport:app');
+        if (! app) return; // webgl not available
 
         container = new pc.Entity(app);
         app.root.addChild(container);
